app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
neo4j